fomo refactor: clean up trainer and test codebase, guard 96/192/224 training
