+++ tests/unit/test_fomo.py
@@ -1,15 +1,11 @@
-"""Unit and end-to-end tests for LibreFOMO."""
+"""Unit tests for LibreFOMO core architecture and registration."""
 
 from __future__ import annotations
 
-import math
 from pathlib import Path
-from typing import Any, List
 
-import numpy as np
 import pytest
 import torch
-from PIL import Image
 
 pytestmark = [pytest.mark.unit, pytest.mark.fomo]
 
@@ -17,34 +13,6 @@ pytestmark = [pytest.mark.unit, pytest.mark.fomo]
 # ===========================================================================
 # Helpers
 # ===========================================================================
-
-
-def _write_tiny_point_dataset(root: Path, imgsz: int = 96) -> Path:
-    """Write a minimal YOLO-format point dataset (1 image, 1 label for train and val splits)."""
-    for split in ("train", "val"):
-        img_dir = root / "images" / split
-        lbl_dir = root / "labels" / split
-        img_dir.mkdir(parents=True, exist_ok=True)
-        lbl_dir.mkdir(parents=True, exist_ok=True)
-        Image.new("RGB", (imgsz, imgsz), color=(128, 64, 32)).save(img_dir / "sample.jpg")
-        (lbl_dir / "sample.txt").write_text("0 0.5 0.5 0.05 0.05\n", encoding="utf-8")
-
-    import yaml
-
-    data_yaml = root / "data.yaml"
-    data_yaml.write_text(
-        yaml.safe_dump(
-            {
-                "path": str(root).replace("\\", "/"),
-                "train": "images/train",
-                "val": "images/val",
-                "nc": 1,
-                "names": {0: "object"},
-            }
-        ),
-        encoding="utf-8",
-    )
-    return data_yaml
 
 
 def _make_random_fomo(size: str = "s", nc: int = 1) -> "LibreFOMO":
@@ -228,6 +196,36 @@ class TestLibreFOMODetectNbClasses:
         assert LibreFOMO.detect_nb_classes(sd_ddp) == 3
 
 
+class TestLibreFOMOFactory:
+    def test_factory_loads_fomo_model(self, tmp_path: Path) -> None:
+        """Verify that LibreYOLO factory auto-detects and loads LibreFOMO from a checkpoint."""
+        from libreyolo import LibreYOLO
+        from libreyolo.models.fomo.nn import LibreFOMOModel
+        from libreyolo.utils.serialization import wrap_libreyolo_checkpoint
+
+        model_module = LibreFOMOModel(size="s", nc=1)
+        state_dict = model_module.state_dict()
+
+        wrapped_ckpt = wrap_libreyolo_checkpoint(
+            state_dict=state_dict,
+            model_family="fomo",
+            size="s",
+            nc=1,
+            names={0: "point"},
+            task="point",
+        )
+
+        weights_file = tmp_path / "LibreFOMOs.pt"
+        torch.save(wrapped_ckpt, weights_file)
+
+        model = LibreYOLO(str(weights_file))
+
+        assert model.family == "fomo"
+        assert model.size == "s"
+        assert model.task == "point"
+        assert model.nb_classes == 1
+
+
 class TestLibreFOMORandomInit:
     """Random-weight models instantiate without errors and have correct defaults."""
 
@@ -246,249 +244,6 @@ class TestLibreFOMORandomInit:
         model = _make_random_fomo()
         assert isinstance(model.model, torch.nn.Module)
 
-class TestLibreFOMOPostprocess:
-    """_postprocess satisfies the PointValidator contract."""
-
-    def test_empty_output_returns_zero_detections(self) -> None:
-        model = _make_random_fomo(size="s", nc=1)
-        from libreyolo.models.fomo.nn import CONFIGS
-
-        hw = CONFIGS["s"]["imgsz"] // 8
-        # All-zero logits → softmax dominated by background → no foreground above threshold
-        logits = torch.zeros(1, 2, hw, hw)
-        result = model._postprocess(
-            logits,
-            conf_thres=0.5,
-            iou_thres=0.45,
-            original_size=(96, 96),
-        )
-        assert result["points"].shape == (0, 4)
-
-    def test_high_confidence_foreground_detected(self) -> None:
-        model = _make_random_fomo(size="s", nc=1)
-        from libreyolo.models.fomo.nn import CONFIGS
-
-        hw = CONFIGS["s"]["imgsz"] // 8
-        # Large positive foreground logit at cell (2, 3)
-        logits = torch.zeros(1, 2, hw, hw)
-        logits[0, 1, 2, 3] = 20.0  # channel 1 = class 0 foreground
-        result = model._postprocess(
-            logits,
-            conf_thres=0.01,
-            iou_thres=0.45,
-            original_size=(96, 96),
-        )
-        assert len(result["points"]) >= 1
-        assert result["points"].shape[1] == 4
-        # class id should be 0 (channel 1 - 1)
-        assert int(result["points"][0, 2].item()) == 0
-
-    def test_postprocess_scales_to_original_size(self) -> None:
-        """Points must be in original-image pixel space, not grid space."""
-        model = _make_random_fomo(size="s", nc=1)
-        from libreyolo.models.fomo.nn import CONFIGS
-
-        hw = CONFIGS["s"]["imgsz"] // 8
-        logits = torch.zeros(1, 2, hw, hw)
-        logits[0, 1, 0, 0] = 20.0  # top-left cell
-        result = model._postprocess(
-            logits,
-            conf_thres=0.01,
-            iou_thres=0.45,
-            original_size=(480, 640),
-        )
-        if len(result["points"]) > 0:
-            x, y = result["points"][0, :2]
-            # scaled pixel should be > 0 and fit in original image
-            assert x.item() >= 0.0
-            assert y.item() >= 0.0
-            assert x.item() <= 640
-            assert y.item() <= 480
-
-    def test_postprocess_output_structure_n_4(self) -> None:
-        """Verify that postprocess returns (N, 4) points with structure [x, y, class, confidence]."""
-        model = _make_random_fomo(size="s", nc=1)
-        from libreyolo.models.fomo.nn import CONFIGS
-
-        hw = CONFIGS["s"]["imgsz"] // 8
-        logits = torch.zeros(1, 2, hw, hw)
-        logits[0, 1, 2, 3] = 20.0  # channel 1 = class 0 foreground
-        result = model._postprocess(
-            logits,
-            conf_thres=0.8,
-            iou_thres=0.45,
-            original_size=(96, 96),
-        )
-        pts = result["points"]
-        assert len(pts) == 1
-        assert pts.shape == (1, 4)
-        # Check structure: [x, y, class, confidence]
-        # Coordinates should be scaled center of cell (2,3)
-        # grid space: x=3, y=2
-        # scaled: x = (3 + 0.5) * (96 / 12) = 3.5 * 8 = 28.0
-        #         y = (2 + 0.5) * (96 / 12) = 2.5 * 8 = 20.0
-        # class: 0.0, confidence: ~1.0
-        np.testing.assert_allclose(pts[0].cpu().numpy(), [28.0, 20.0, 0.0, 1.0], atol=1e-3)
-
-    def test_inference_returns_points_object_with_n_4(self) -> None:
-        """Verify that calling predict() on FOMO returns Results with a Points payload of shape (N, 4)."""
-        model = _make_random_fomo(size="s", nc=1)
-        from libreyolo.models.fomo.nn import CONFIGS
-
-        hw = CONFIGS["s"]["imgsz"] // 8
-        # Mock _forward to return a logit with a peak at (2, 3)
-        def _mock_forward(x):
-            logits = torch.zeros(x.shape[0], 2, hw, hw)
-            logits[:, 1, 2, 3] = 20.0
-            return logits
-
-        model._forward = _mock_forward
-
-        img = Image.new("RGB", (96, 96))
-        results = model.predict(img, conf=0.8)
-
-        if isinstance(results, list):
-            results = results[0]
-
-        assert results.points is not None
-        assert results.points.data.shape == (1, 4)
-        np.testing.assert_allclose(results.points.xy[0].cpu().numpy(), [28.0, 20.0], atol=1e-3)
-        assert int(results.points.cls[0].item()) == 0
-        assert results.points.conf[0].item() > 0.99
-
-
-
-class TestLibreFOMOValPreprocessor:
-    """Validate FOMOValPreprocessor output shape, dtype, and pixel range."""
-
-    def _make_preprocessor(self, imgsz: int = 96):
-        from libreyolo.validation.preprocessors import FOMOValPreprocessor
-
-        return FOMOValPreprocessor(img_size=(imgsz, imgsz))
-
-    def test_output_shape(self) -> None:
-        preproc = self._make_preprocessor(96)
-        img = np.random.randint(0, 256, (200, 300, 3), dtype=np.uint8)
-        targets = np.zeros((0, 5), dtype=np.float32)
-        chw, _ = preproc(img, targets, (96, 96))
-        assert chw.shape == (3, 96, 96)
-
-    def test_output_dtype_float32(self) -> None:
-        preproc = self._make_preprocessor(96)
-        img = np.ones((96, 96, 3), dtype=np.uint8) * 128
-        chw, _ = preproc(img, np.zeros((0, 5), dtype=np.float32), (96, 96))
-        assert chw.dtype == np.float32
-
-    def test_output_range_minus1_to_1(self) -> None:
-        preproc = self._make_preprocessor(96)
-        # Black image → pixel 0 → (0/255 - 0.5)/0.5 = -1.0
-        img_black = np.zeros((96, 96, 3), dtype=np.uint8)
-        chw_black, _ = preproc(img_black, np.zeros((0, 5), dtype=np.float32), (96, 96))
-        assert chw_black.min() == pytest.approx(-1.0, abs=1e-4)
-        # White image → pixel 255 → (1.0 - 0.5)/0.5 = 1.0
-        img_white = np.full((96, 96, 3), 255, dtype=np.uint8)
-        chw_white, _ = preproc(img_white, np.zeros((0, 5), dtype=np.float32), (96, 96))
-        assert chw_white.max() == pytest.approx(1.0, abs=1e-4)
-
-    def test_custom_normalization_flag(self) -> None:
-        from libreyolo.validation.preprocessors import FOMOValPreprocessor
-
-        p = FOMOValPreprocessor(img_size=(96, 96))
-        assert p.custom_normalization is True
-
-    def test_wants_unresized_image_flag(self) -> None:
-        from libreyolo.validation.preprocessors import FOMOValPreprocessor
-
-        p = FOMOValPreprocessor(img_size=(96, 96))
-        assert p.wants_unresized_image is True
-
-    def test_uses_letterbox_false(self) -> None:
-        from libreyolo.validation.preprocessors import FOMOValPreprocessor
-
-        p = FOMOValPreprocessor(img_size=(96, 96))
-        assert p.uses_letterbox is False
-
-    def test_target_scaling(self) -> None:
-        """Box coordinates must be scaled proportionally when resizing."""
-        preproc = self._make_preprocessor(96)
-        img = np.zeros((200, 400, 3), dtype=np.uint8)  # orig 200×400
-        # Box: [x1, y1, x2, y2, cls] in original pixels (letterboxed from dataset)
-        targets = np.array([[100.0, 50.0, 200.0, 100.0, 0.0]], dtype=np.float32)
-        _, padded_targets = preproc(img, targets, (96, 96))
-        # scale_x = 96 / 400 = 0.24 ; scale_y = 96 / 200 = 0.48
-        np.testing.assert_allclose(padded_targets[0, 0], 100.0 * (96 / 400), atol=1e-4)
-        np.testing.assert_allclose(padded_targets[0, 1], 50.0 * (96 / 200), atol=1e-4)
-
-    def test_empty_targets_padded_to_max_labels(self) -> None:
-        preproc = self._make_preprocessor(96)
-        img = np.zeros((96, 96, 3), dtype=np.uint8)
-        _, padded = preproc(img, np.zeros((0, 5), dtype=np.float32), (96, 96))
-        assert padded.shape == (preproc.max_labels, 5)
-        assert padded.sum() == 0.0
-
-
-class TestFOMOTrainTransform:
-    """Verify that FOMOTrainTransform performs direct stretch resizing and scales coordinates."""
-
-    def test_stretch_resize_and_coordinate_scaling(self) -> None:
-        from libreyolo.models.fomo.dataset import FOMOTrainTransform
-
-        img = np.zeros((200, 400, 3), dtype=np.uint8)
-        targets = np.array([[150.0, 50.0, 250.0, 150.0, 0.0]], dtype=np.float32)
-
-        transform = FOMOTrainTransform(max_labels=10, flip_prob=0.0, hsv_prob=0.0)
-        assert transform.wants_unresized_image is True
-
-        img_out, targets_out = transform(img, targets, (96, 96))
-
-        assert img_out.shape == (3, 96, 96)
-
-        expected = np.array([0.0, 48.0, 48.0, 24.0, 48.0], dtype=np.float32)
-        np.testing.assert_allclose(targets_out[0], expected, atol=1e-4)
-
-    def test_empty_targets(self) -> None:
-        from libreyolo.models.fomo.dataset import FOMOTrainTransform
-
-        img = np.zeros((100, 200, 3), dtype=np.uint8)
-        targets = np.zeros((0, 5), dtype=np.float32)
-
-        transform = FOMOTrainTransform(max_labels=10, flip_prob=0.0, hsv_prob=0.0)
-        img_out, targets_out = transform(img, targets, (96, 96))
-
-        assert img_out.shape == (3, 96, 96)
-        assert targets_out.shape == (10, 5)
-        assert targets_out.sum() == 0.0
-
-
-class TestLibreFOMOParseGtPoints:
-    """_parse_gt_points must delegate to the validator's box-centre helper."""
-
-    def test_delegates_to_validator_helper(self) -> None:
-        from libreyolo.validation.point_validator import PointValidator
-
-        model = _make_random_fomo(size="s", nc=1)
-
-        # Build a minimal validator scaffold
-        validator = object.__new__(PointValidator)
-        from libreyolo.validation.point_validator import _DEFAULT_DIST_THRESHOLDS
-
-        validator._dist_thresholds = _DEFAULT_DIST_THRESHOLDS
-        validator._primary_threshold = _DEFAULT_DIST_THRESHOLDS[0]
-        validator._records = []
-        validator.nc = 1
-        validator._actual_imgsz = 96
-        validator.config = type("_Cfg", (), {"verbose": False})()
-        validator.seen = 0
-        validator.val_preproc = type("_Preproc", (), {"uses_letterbox": False})()
-        validator.model = model
-
-        gt_row = np.array([[43.2, 43.2, 52.8, 52.8, 0.0]], dtype=np.float32)
-        xy, cls = model._parse_gt_points(gt_row, orig_h=96, orig_w=96, validator=validator)
-
-        assert xy.shape == (1, 2)
-        np.testing.assert_allclose(xy[0], [0.5, 0.5], atol=1e-5)
-        assert cls[0] == 0
-
 
 class TestLibreFOMODownloadURL:
     def test_valid_filename_returns_url(self) -> None:
@@ -502,463 +257,3 @@ class TestLibreFOMODownloadURL:
 
         url = LibreFOMO.get_download_url("SomeOtherModel.pt")
         assert url is None
-
-
-# ===========================================================================
-# End-to-end: PointValidator integration
-# ===========================================================================
-
-
-class TestLibreFOMOEndToEnd:
-    """Run a full validation pass using PointValidator on a synthetic dataset."""
-
-    def test_point_validator_integration(self, tmp_path: Path) -> None:
-        from libreyolo.validation import PointValidator, ValidationConfig
-
-        data_yaml = _write_tiny_point_dataset(tmp_path, imgsz=96)
-
-        # Build a random-weight FOMO model that always predicts one foreground
-        # point at the centre of the grid.
-        model = _make_random_fomo(size="s", nc=1)
-
-        # Monkey-patch _postprocess so we can inspect the PointValidator pipeline
-        # without needing trained weights.
-        def _fixed_postprocess(output, conf_thres, iou_thres, original_size, **kwargs):
-            w, h = original_size
-            return {
-                "points": torch.tensor([[w / 2.0, h / 2.0, 0.0, 0.9]], dtype=torch.float32),
-            }
-
-        model._postprocess = _fixed_postprocess
-
-        config = ValidationConfig(
-            data=str(data_yaml),
-            batch_size=1,
-            imgsz=96,
-            num_workers=0,
-            verbose=False,
-            device="cpu",
-            save_dir=str(tmp_path / "val_out"),
-        )
-        validator = PointValidator(model, config)
-        metrics = validator.run()
-
-        # The validator ran without error
-        assert "metrics/precision" in metrics
-        assert "metrics/recall" in metrics
-        assert "metrics/f1" in metrics
-        assert "speed/images_seen" in metrics
-        assert metrics["speed/images_seen"] == 1
-
-    def test_training_checkpoint_reload_fallback(self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
-        """Verify that train() correctly falls back to last_checkpoint if best_checkpoint is missing/empty."""
-        from libreyolo.models.fomo.model import LibreFOMO
-        
-        dummy_ckpt = tmp_path / "last.pt"
-        model = _make_random_fomo(size="s", nc=1)
-        from libreyolo.utils.serialization import wrap_libreyolo_checkpoint
-        ckpt_dict = wrap_libreyolo_checkpoint(
-            model.model.state_dict(),
-            model_family="fomo",
-            size="s",
-            task="point",
-            nc=1,
-        )
-        torch.save(ckpt_dict, dummy_ckpt)
-        
-        class DummyTrainer:
-            def __init__(self, *args: Any, **kwargs: Any) -> None:
-                pass
-            def train(self) -> dict[str, Any]:
-                return {
-                    "epoch_losses": [0.5],
-                    "final_loss": 0.5,
-                    "best_checkpoint": "",
-                    "last_checkpoint": str(dummy_ckpt),
-                }
-                
-        monkeypatch.setattr("libreyolo.models.fomo.trainer.FOMOTrainer", DummyTrainer)
-        
-        loaded_path = None
-        def mock_load_weights(path: str | Path) -> None:
-            nonlocal loaded_path
-            loaded_path = path
-            
-        monkeypatch.setattr(model, "_load_weights", mock_load_weights)
-        
-        results = model.train("dummy.yaml", allow_experimental=True)
-        assert loaded_path == str(dummy_ckpt)
-
-    def test_train_resume_raises_when_no_checkpoint(self) -> None:
-        model = _make_random_fomo(size="s", nc=1)
-        with pytest.raises(ValueError, match="resume=True requires a checkpoint"):
-            model.train("dummy.yaml", allow_experimental=True, resume=True)
-
-    def test_train_resume_calls_trainer_resume(self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
-        model = _make_random_fomo(size="s", nc=1)
-        dummy_ckpt = tmp_path / "last.pt"
-        model.model_path = dummy_ckpt
-
-        setup_called = False
-        resume_called_with = None
-
-        class DummyTrainer:
-            def __init__(self, *args: Any, **kwargs: Any) -> None:
-                pass
-            def setup(self) -> None:
-                nonlocal setup_called
-                setup_called = True
-            def resume(self, path: str) -> None:
-                nonlocal resume_called_with
-                resume_called_with = path
-            def train(self) -> dict[str, Any]:
-                return {"epoch_losses": [0.5], "final_loss": 0.5}
-
-        monkeypatch.setattr("libreyolo.models.fomo.trainer.FOMOTrainer", DummyTrainer)
-
-        results = model.train("dummy.yaml", allow_experimental=True, resume=True)
-        assert setup_called is True
-        assert resume_called_with == str(dummy_ckpt)
-
-    def test_train_seeds_rngs(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Verify that model.train seeds random, numpy, and torch before trainer construction."""
-        model = _make_random_fomo(size="s", nc=1)
-        
-        seeds_recorded = []
-        
-        class DummyTrainer:
-            def __init__(self, *args: Any, **kwargs: Any) -> None:
-                import random
-                import numpy as np
-                import torch
-                seeds_recorded.append((
-                    random.randint(0, 1000000),
-                    np.random.randint(0, 1000000),
-                    torch.randint(0, 1000000, (1,)).item()
-                ))
-
-            def train(self) -> dict[str, Any]:
-                return {"epoch_losses": [0.5], "final_loss": 0.5}
-
-        monkeypatch.setattr("libreyolo.models.fomo.trainer.FOMOTrainer", DummyTrainer)
-
-        model.train("dummy.yaml", allow_experimental=True, seed=0)
-        state_0_first = seeds_recorded[0]
-
-        seeds_recorded.clear()
-        model.train("dummy.yaml", allow_experimental=True, seed=0)
-        state_0_second = seeds_recorded[0]
-
-        seeds_recorded.clear()
-        model.train("dummy.yaml", allow_experimental=True, seed=100)
-        state_100 = seeds_recorded[0]
-
-        assert state_0_first == state_0_second
-        assert state_0_first != state_100
-
-    def test_trainer_build_yolo_datasets_omitted_label_files(self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
-        """Verify that _build_yolo_datasets correctly resolves label files when image files are pre-resolved but label files are None/omitted."""
-        from libreyolo.models.fomo.trainer import FOMOTrainer
-        
-        # Mock load_data_config
-        def mock_load_data_config(*args, **kwargs):
-            return {
-                "root": str(tmp_path),
-                "train_img_files": [tmp_path / "images" / "train" / "img1.jpg"],
-                "train_label_files": None,
-                "val_img_files": [tmp_path / "images" / "val" / "img2.jpg"],
-                "val_label_files": None,
-                "nc": 1,
-                "names": {0: "object"},
-            }
-        
-        img1 = tmp_path / "images" / "train" / "img1.jpg"
-        img2 = tmp_path / "images" / "val" / "img2.jpg"
-        img1.parent.mkdir(parents=True, exist_ok=True)
-        img2.parent.mkdir(parents=True, exist_ok=True)
-        img1.touch()
-        img2.touch()
-        
-        lbl1 = tmp_path / "labels" / "train" / "img1.txt"
-        lbl2 = tmp_path / "labels" / "val" / "img2.txt"
-        lbl1.parent.mkdir(parents=True, exist_ok=True)
-        lbl2.parent.mkdir(parents=True, exist_ok=True)
-        lbl1.touch()
-        lbl2.touch()
-        
-        import libreyolo.data as data_mod
-        monkeypatch.setattr(data_mod, "load_data_config", mock_load_data_config)
-        
-        model = _make_random_fomo(size="s", nc=1)
-        trainer = FOMOTrainer(
-            model=model.model,
-            wrapper_model=model,
-            data="dummy.yaml",
-            epochs=1,
-            batch=2,
-            imgsz=96,
-            device="cpu",
-            project=str(tmp_path / "runs"),
-            workers=0,
-        )
-        
-        train_ds, val_ds = trainer._build_yolo_datasets(input_size=96, grid_size=12)
-        assert train_ds.label_files is not None
-        assert len(train_ds.label_files) == 1
-        assert val_ds.label_files is not None
-        assert len(val_ds.label_files) == 1
-
-    def test_trainer_class_count_mismatch_fails_fast(self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
-        """Verify that _build_yolo_datasets raises a ValueError if dataset nc != model nc and wrapper_model is None."""
-        from libreyolo.models.fomo.trainer import FOMOTrainer
-        
-        # Mock load_data_config
-        def mock_load_data_config(*args, **kwargs):
-            return {
-                "root": str(tmp_path),
-                "train_img_files": [tmp_path / "images" / "train" / "img1.jpg"],
-                "train_label_files": None,
-                "val_img_files": [tmp_path / "images" / "val" / "img2.jpg"],
-                "val_label_files": None,
-                "nc": 2,  # Mismatch with model nc=1
-                "names": {0: "class1", 1: "class2"},
-            }
-            
-        img1 = tmp_path / "images" / "train" / "img1.jpg"
-        img2 = tmp_path / "images" / "val" / "img2.jpg"
-        img1.parent.mkdir(parents=True, exist_ok=True)
-        img2.parent.mkdir(parents=True, exist_ok=True)
-        img1.touch()
-        img2.touch()
-        
-        lbl1 = tmp_path / "labels" / "train" / "img1.txt"
-        lbl2 = tmp_path / "labels" / "val" / "img2.txt"
-        lbl1.parent.mkdir(parents=True, exist_ok=True)
-        lbl2.parent.mkdir(parents=True, exist_ok=True)
-        lbl1.touch()
-        lbl2.touch()
-        
-        import libreyolo.data as data_mod
-        monkeypatch.setattr(data_mod, "load_data_config", mock_load_data_config)
-        
-        model = _make_random_fomo(size="s", nc=1)
-        trainer = FOMOTrainer(
-            model=model.model,
-            wrapper_model=None,  # None to trigger the error path
-            data="dummy.yaml",
-            epochs=1,
-            batch=2,
-            imgsz=96,
-            device="cpu",
-            project=str(tmp_path / "runs"),
-            workers=0,
-        )
-        
-        with pytest.raises(ValueError, match="wrapper_model is None — cannot rebuild head safely"):
-            trainer._build_yolo_datasets(input_size=96, grid_size=12)
-
-    def test_val_dispatch_uses_point_validator(self) -> None:
-        """BaseModel.val() must route task='point' to PointValidator."""
-        from libreyolo.models.base.model import BaseModel
-
-        model = _make_random_fomo(size="s", nc=1)
-        assert model.task == "point"
-
-        # Augmented validation must be blocked for point models
-        with pytest.raises(ValueError, match="Augmented validation"):
-            BaseModel.val(model, data="unused.yaml", imgsz=96, augment=True)
-
-    def test_train_without_flag_raises(self) -> None:
-        model = _make_random_fomo(size="s")
-        with pytest.raises(NotImplementedError, match="allow_experimental"):
-            model.train("dummy.yaml")
-
-class TestFOMODatasets:
-    """Verify FOMOYOLODataset and FOMOAugmentedDataset functionality on synthetic data."""
-
-    def test_fomo_yolo_dataset(self, tmp_path: Path) -> None:
-        from libreyolo.models.fomo.dataset import FOMOYOLODataset
-
-        # Set up synthetic image and label
-        img_dir = tmp_path / "images"
-        lbl_dir = tmp_path / "labels"
-        img_dir.mkdir(parents=True)
-        lbl_dir.mkdir(parents=True)
-
-        img_path = img_dir / "sample.jpg"
-        lbl_path = lbl_dir / "sample.txt"
-
-        # 96x96 synthetic image
-        Image.new("RGB", (96, 96), color=(255, 128, 64)).save(img_path)
-        # YOLO box format: cls cx cy w h
-        # One point at cx=0.5, cy=0.5
-        lbl_path.write_text("0 0.5 0.5 0.1 0.1\n", encoding="utf-8")
-
-        dataset = FOMOYOLODataset(
-            img_files=[img_path],
-            label_files=[lbl_path],
-            input_size=96,
-            grid_size=12,
-        )
-
-        assert len(dataset) == 1
-        img_tensor, grid, img_info, idx = dataset[0]
-
-        # Check shapes and types
-        assert img_tensor.shape == (3, 96, 96)
-        assert img_tensor.dtype == torch.float32
-        assert grid.shape == (12, 12)
-        assert grid.dtype == torch.long
-        assert img_info == (96, 96)
-        assert idx == 0
-
-        # Center should be class 0 + 1 = 1
-        assert grid[6, 6] == 1
-        assert grid.sum() == 1
-
-    def test_fomo_yolo_dataset_polygon_labels(self, tmp_path: Path) -> None:
-        from libreyolo.models.fomo.dataset import FOMOYOLODataset
-
-        # Set up synthetic image and label
-        img_dir = tmp_path / "images"
-        lbl_dir = tmp_path / "labels"
-        img_dir.mkdir(parents=True, exist_ok=True)
-        lbl_dir.mkdir(parents=True, exist_ok=True)
-
-        img_path = img_dir / "sample_poly.jpg"
-        lbl_path = lbl_dir / "sample_poly.txt"
-
-        Image.new("RGB", (96, 96), color=(255, 128, 64)).save(img_path)
-        lbl_path.write_text("0 0.4 0.4 0.6 0.4 0.6 0.6 0.4 0.6\n", encoding="utf-8")
-
-        dataset = FOMOYOLODataset(
-            img_files=[img_path],
-            label_files=[lbl_path],
-            input_size=96,
-            grid_size=12,
-        )
-
-        assert len(dataset) == 1
-        _, grid, _, _ = dataset[0]
-
-        assert grid[6, 6] == 1
-        assert grid.sum() == 1
-
-    def test_fomo_augmented_dataset(self) -> None:
-        from libreyolo.models.fomo.dataset import FOMOAugmentedDataset
-
-        # Create a mock augmented dataset that returns (img, targets, img_info, img_id)
-        # YOLOX-style image shape (C, H, W) in BGR
-        mock_img = np.zeros((3, 96, 96), dtype=np.uint8)
-        # target row: [class, cx, cy, w, h, ...]
-        mock_targets = np.array([
-            [0.0, 48.0, 48.0, 10.0, 10.0]
-        ], dtype=np.float32)
-        mock_img_info = (96, 96)
-        mock_img_id = 42
-
-        class MockBaseDataset:
-            def __len__(self):
-                return 1
-            def __getitem__(self, idx):
-                return mock_img, mock_targets, mock_img_info, mock_img_id
-            def close_mosaic(self):
-                self.closed = True
-
-        base_ds = MockBaseDataset()
-        base_ds.closed = False
-
-        dataset = FOMOAugmentedDataset(
-            augmented_dataset=base_ds,
-            input_size=96,
-            grid_size=12,
-        )
-
-        assert len(dataset) == 1
-        img_tensor, grid, img_info, img_id = dataset[0]
-
-        assert img_tensor.shape == (3, 96, 96)
-        assert grid.shape == (12, 12)
-        assert img_info == (96, 96)
-        assert img_id == 42
-
-        assert grid[6, 6] == 1
-        assert grid.sum() == 1
-
-        dataset.close_mosaic()
-        assert base_ds.closed is True
-
-
-# ===========================================================================
-# Decode helper unit tests
-# ===========================================================================
-
-
-class TestDecodePoints:
-    """Unit tests for _decode_points (the internal NMS peak-finder)."""
-
-    def test_all_background_returns_empty(self) -> None:
-        from libreyolo.models.fomo.utils import decode_points_from_logits as _decode_points
-
-        # All logits zero → softmax gives equal mass to bg and fg → obj_probs ~0.5
-        # depending on nc. With threshold 0.6 nothing should fire.
-        logits = torch.zeros(1, 2, 6, 6)  # nc=1
-        results = _decode_points(logits, conf_threshold=0.6)
-        assert results[0].shape[0] == 0
-
-    def test_peak_at_known_location(self) -> None:
-        from libreyolo.models.fomo.utils import decode_points_from_logits as _decode_points
-
-        logits = torch.zeros(1, 2, 6, 6)
-        logits[0, 1, 3, 4] = 20.0  # large foreground logit at (y=3, x=4)
-        results = _decode_points(logits, conf_threshold=0.01)
-        assert len(results[0]) >= 1
-        # x should be 4, y should be 3
-        assert int(results[0][0, 0].item()) == 4  # x
-        assert int(results[0][0, 1].item()) == 3  # y
-
-    def test_nms_radius_suppresses_neighbours(self) -> None:
-        from libreyolo.models.fomo.utils import decode_points_from_logits as _decode_points
-
-        logits = torch.zeros(1, 2, 6, 6)
-        # Two adjacent peaks; with nms_radius=2 the second should be suppressed
-        logits[0, 1, 3, 3] = 20.0
-        logits[0, 1, 3, 4] = 18.0  # within radius 2 of (3,3)
-        results_r2 = _decode_points(logits, conf_threshold=0.01, nms_radius=2)
-        results_r0 = _decode_points(logits, conf_threshold=0.01, nms_radius=0)
-        assert len(results_r2[0]) < len(results_r0[0])
-
-    def test_class_index_correct(self) -> None:
-        """Returned class channel index must equal the argmax foreground channel."""
-        from libreyolo.models.fomo.utils import decode_points_from_logits as _decode_points
-
-        logits = torch.zeros(1, 3, 4, 4)  # nc=2
-        logits[0, 2, 1, 1] = 20.0  # channel 2 → 0-based class id 1
-        results = _decode_points(logits, conf_threshold=0.01)
-        assert len(results[0]) >= 1
-        cls_channel = int(results[0][0, 2].item())  # raw channel index
-        assert cls_channel == 2
-
-    def test_batch_dimension_respected(self) -> None:
-        from libreyolo.models.fomo.utils import decode_points_from_logits as _decode_points
-
-        # nc=1: softmax gives probs 0.5 each when all logits are zero.
-        # Use threshold > 0.5 so only the explicitly boosted cells fire.
-        logits = torch.zeros(3, 2, 4, 4)
-        logits[0, 1, 0, 0] = 20.0   # batch 0: large foreground at (0,0)
-        logits[2, 1, 3, 3] = 20.0   # batch 2: large foreground at (3,3)
-        # batch 1 stays all-zeros → obj_probs ≈ 0.5, below threshold 0.6
-        results = _decode_points(logits, conf_threshold=0.6)
-        assert len(results) == 3
-        assert len(results[0]) >= 1
-        assert len(results[1]) == 0
-        assert len(results[2]) >= 1
-
-
-    def test_max_points_limit(self) -> None:
-        from libreyolo.models.fomo.utils import decode_points_from_logits as _decode_points
-
-        logits = torch.zeros(1, 2, 8, 8)
-        logits[0, 1, :, :] = 20.0  # every cell fires
-        results = _decode_points(logits, conf_threshold=0.01, nms_radius=0, max_points=3)
-        assert len(results[0]) <= 3
